avances generales

+++ Metamapa/M-Agregador-Service/src/main/java/Agregador/business/Colecciones/CriterioFuenteDeDatos.java
@@ -3,17 +3,19 @@ package Agregador.business.Colecciones;
 import lombok.Getter;
 import Agregador.business.Hechos.Hecho;
 
+import java.math.BigInteger;
+
 public class CriterioFuenteDeDatos implements Criterio {
   @Getter
   private Integer idFuenteDeDatos;
 
-  public CriterioFuenteDeDatos(Integer idFuenteDeDatos)
-  {
+  public CriterioFuenteDeDatos(Integer idFuenteDeDatos) {
     this.idFuenteDeDatos = idFuenteDeDatos;
   }
+
   @Override
   public boolean cumple(Hecho hechoAValidar){
-    Integer idFuenteDeDatosAValidar = hechoAValidar.getId();
+    BigInteger idFuenteDeDatosAValidar = hechoAValidar.getId();
 
     return this.getIdFuenteDeDatos().equals(idFuenteDeDatosAValidar);
   }
